+++ packages/at_stream/lib/at_stream.dart
@@ -1,0 +1,9 @@
+/// # at_stream
+///
+/// See the examples for usage patterns.
+library;
+
+import 'package:meta/meta.dart' show experimental;
+
+@experimental
+export 'src/at_notification_stream_channel.dart';