+++ packages/at_stream/lib/sources.dart
@@ -1,0 +1,14 @@
+/// Connect various sources into stream channels
+/// - Socket
+/// - Stdio (stdin, stdout)
+/// - UDS (Unix domain sockets)
+library;
+
+import 'package:meta/meta.dart' show experimental;
+
+@experimental
+export 'src/stream_channel_sources/stream_channel_stdio.dart';
+@experimental
+export 'src/stream_channel_sources/stream_channel_socket.dart';
+@experimental
+export 'src/stream_channel_sources/stream_channel_uds.dart';