+++ packages/at_stream/lib/src/stream_channel_sources/stream_channel_socket.dart
@@ -1,0 +1,17 @@
+import 'dart:io' show Socket;
+import 'dart:typed_data' show Uint8List;
+
+import 'package:at_stream/src/transformers/common_transformers.dart'
+    show CommonTransformers;
+import 'package:stream_channel/stream_channel.dart';
+
+// Turns a Socket which is a Stream<Uint8List>, StreamSink<List<int>>
+// into a StreamChannel<Uint8List>
+extension StreamChannelSocket on Socket {
+  StreamChannel<Uint8List> toStreamChannel() {
+    return StreamChannel<Uint8List>(
+      this,
+      CommonTransformers.listIntToUint8ListSink.bind(this),
+    );
+  }
+}