+++ packages/at_stream/lib/transformers.dart
@@ -1,0 +1,14 @@
+/// Common transformers and transformation functions
+///
+/// - Provides the .join function for Stream, StreamSink, StreamChannel
+/// - Provides CommonTransformers
+///   (Currently only [List<int>] to [Uint8List] for Stream/StreamSink)
+library;
+
+import 'package:meta/meta.dart' show experimental;
+
+@experimental
+export 'src/transformers/common_transformers.dart';
+
+@experimental
+export 'src/transformers/transformer_join_extensions.dart';